backend/requirements.txt
asgiref==3.8.1
aStarEngine_package==1.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
idna==3.10
joblib==1.4.2
networkx==3.4.2
numpy==2.1.3
osmapi==4.2.0
osmnx==2.0.0
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.10
pyogrio==0.10.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
six==1.17.0
sqlparse==0.5.2
threadpoolctl==3.5.0
tzdata==2024.2
urllib3==2.2.3
djangorestframework-simplejwt
